+++ src/handlers/coverage.ts
@@ -9,3 +9,13 @@ export const getPullRequestGitDiffHandler = async (args: any) => {
     pullRequestNumber
   );
 };
+
+export const getRepositoryPullRequestFilesCoverageHandler = async (args: any) => {
+  const { provider, organization, repository, pullRequestNumber } = args;
+  return await CoverageService.getRepositoryPullRequestFilesCoverage(
+    provider,
+    organization,
+    repository,
+    pullRequestNumber
+  );
+};
