--- src/handlers/getFileCoverageHandler.ts
@@ -1,7 +1,0 @@
-import { RepositoryService } from '../api/client/index.js';
-
-export const getFileCoverageHandler = async (args: any) => {
-  const { provider, organization, repository, fileId } = args;
-
-  return await RepositoryService.getFileCoverage(provider, organization, repository, fileId);
-};
