+++ src/handlers/index.ts
@@ -1,11 +1,5 @@
-export * from './getFileCoverageHandler.js';
-export * from './getFileIssuesHandler.js';
-export * from './getPullRequestGitDiffHandler.js';
-export * from './getRepositoryPullRequestFilesCoverageHandler.js';
-export * from './listFilesHandler.js';
-export * from './listRepositoryPullRequestsHandler.js';
-export * from './listOrganizationRepositoriesHandler.js';
-export * from './listPullRequestIssuesHandler.js';
-export * from './searchRepositoryIssuesHandler.js';
-export * from './searchSecurityItemsHandler.js';
-export * from './listOrganizationRepositoriesHandler.js';
+export * from './coverage.js';
+export * from './repository.js';
+export * from './analysis.js';
+export * from './organization.js';
+export * from './security.js';
