+++ src/tools/listFileIssuesTool.ts
@@ -1,13 +1,14 @@
 import { Tool } from '@modelcontextprotocol/sdk/types.js';
-import { repositorySchema } from './utils.js';
+import { defaultPagination, repositorySchema } from './utils.js';
 
-export const getFileCoverageTool: Tool = {
-  name: 'codacy_get_file_coverage',
-  description: 'Get coverage information for a file in the head commit of a repository branch.',
+export const listFileIssuesTool: Tool = {
+  name: 'codacy_get_file_issues',
+  description: 'Get the issue list for a file in a repository',
   inputSchema: {
     type: 'object',
     properties: {
       ...repositorySchema,
+      ...defaultPagination,
       fileId: {
         type: 'string',
         description: 'Identifier of a file in a specific commit',
