+++ src/tools/listFilesTool.ts
@@ -1,4 +1,5 @@
 import { Tool } from '@modelcontextprotocol/sdk/types.js';
+import { getPaginationWithSorting, repositorySchema } from './utils.js';
 
 export const listFilesTool: Tool = {
   name: 'codacy_list_files',
@@ -6,19 +7,10 @@ export const listFilesTool: Tool = {
   inputSchema: {
     type: 'object',
     properties: {
-      provider: {
-        type: 'string',
-        description:
-          "Organization's git provider: GitHub (gh), GitLab (gl) or BitBucket (bb). Accepted values: gh, gl, bb.",
-      },
-      organization: {
-        type: 'string',
-        description: 'Organization name on the Git provider',
-      },
-      repository: {
-        type: 'string',
-        description: 'Repository name on the Git provider organization',
-      },
+      ...repositorySchema,
+      ...getPaginationWithSorting(
+        "Field used to sort the list of files. The allowed values are 'filename', 'issues', 'grade', 'duplication', 'complexity', and 'coverage'."
+      ),
       branch: {
         type: 'string',
         description: 'Name of a repository branch enabled on Codacy',
@@ -26,25 +18,6 @@ export const listFilesTool: Tool = {
       search: {
         type: 'string',
         description: 'Filter files that include this string anywhere in their relative path',
-      },
-      sort: {
-        type: 'string',
-        description:
-          "Field used to sort the list of files. The allowed values are 'filename', 'issues', 'grade', 'duplication', 'complexity', and 'coverage'.",
-      },
-      direction: {
-        type: 'string',
-        description:
-          "Sort direction. Possible values are 'asc' (ascending) or 'desc' (descending).",
-      },
-      cursor: {
-        type: 'string',
-        description: 'Pagination cursor for next page of results',
-      },
-      limit: {
-        type: 'number',
-        description: 'Maximum number of results to return (default 100, max 100)',
-        default: 100,
       },
     },
   },
