+++ src/tools/listPullRequestIssuesTool.ts
@@ -1,4 +1,5 @@
 import { Tool } from '@modelcontextprotocol/sdk/types.js';
+import { defaultPagination, repositorySchema } from './utils.js';
 
 export const listPullRequestIssuesTool: Tool = {
   name: 'codacy_list_pull_request_issues',
@@ -7,19 +8,8 @@ export const listPullRequestIssuesTool: Tool = {
   inputSchema: {
     type: 'object',
     properties: {
-      provider: {
-        type: 'string',
-        description:
-          "Organization's git provider: GitHub (gh), GitLab (gl) or BitBucket (bb). Accepted values: gh, gl, bb.",
-      },
-      organization: {
-        type: 'string',
-        description: 'Organization name on the Git provider',
-      },
-      repository: {
-        type: 'string',
-        description: 'Repository name on the Git provider organization',
-      },
+      ...repositorySchema,
+      ...defaultPagination,
       pullRequestNumber: {
         type: 'number',
         description: 'Pull request number',
@@ -31,15 +21,6 @@ export const listPullRequestIssuesTool: Tool = {
       onlyPotential: {
         type: 'boolean',
         description: 'If true, retrieves only potential issues',
-      },
-      cursor: {
-        type: 'string',
-        description: 'Pagination cursor for next page of results',
-      },
-      limit: {
-        type: 'number',
-        description: 'Maximum number of results to return (default 100, max 100)',
-        default: 100,
       },
     },
   },
